heights int -> long

+++ src/main/java/monero/daemon/model/MoneroBlock.java
@@ -181,7 +181,7 @@ public class MoneroBlock extends MoneroBlockHeader {
   }
   
   @Override
-  public MoneroBlock setHeight(Integer height) {
+  public MoneroBlock setHeight(Long height) {
     super.setHeight(height);
     return this;
   }
